test: add test and rename
