+++ pkg/statistics/handle/autoanalyze/BUILD.bazel
@@ -12,7 +12,7 @@ go_library(
         "//pkg/sessionctx",
         "//pkg/sessionctx/variable",
         "//pkg/statistics",
-        "//pkg/statistics/handle/autoanalyze/util",
+        "//pkg/statistics/handle/autoanalyze/exec",
         "//pkg/statistics/handle/lockstats",
         "//pkg/statistics/handle/logutil",
         "//pkg/statistics/handle/types",
