--- pkg/statistics/handle/autoanalyze/util/BUILD.bazel
@@ -1,21 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "util",
-    srcs = ["util.go"],
-    importpath = "github.com/pingcap/tidb/pkg/statistics/handle/autoanalyze/util",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/metrics",
-        "//pkg/parser/ast",
-        "//pkg/sessionctx",
-        "//pkg/statistics",
-        "//pkg/statistics/handle/logutil",
-        "//pkg/statistics/handle/types",
-        "//pkg/statistics/handle/util",
-        "//pkg/util/chunk",
-        "//pkg/util/sqlescape",
-        "//pkg/util/sqlexec",
-        "@org_uber_go_zap//:zap",
-    ],
-)
